+++ src/ReceivingList/ExportSettingsModal/utils/createExportReport.js
@@ -89,6 +89,9 @@ export const createExportReport = (
       supplement: Boolean(pieceData.supplement),
       displayOnHolding: pieceData.displayOnHolding,
       itemHRID: item?.hrid,
+      receivingStatus: pieceData.receivingStatus,
+      internalNote: pieceData.internalNote,
+      externalNote: pieceData.externalNote,
     };
   };